requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
docopt==0.6.2
idna==3.4
pipreqs==0.4.13
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
urllib3==2.0.7
yarg==0.1.9
